agent/requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-google
livekit-plugins-noise-cancellation
mem0ai
langchain_community
duckduckgo-search
requests
python-dotenv
flask
flask-cors
livekit-api